pyproject.toml
[project]
name = "slack-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.27.0",
]
